chore!: uint32 for blob size (#1181)
Closes https://github.com/celestiaorg/celestia-app/issues/1153

I will follow-up with a PR that updates the `specs-staging` branch

+++ app/process_proposal.go
@@ -107,7 +107,7 @@ func (app *App) ProcessProposal(req abci.RequestProcessProposal) abci.ResponsePr
 				}
 			}
 
-			commitment, err := inclusion.GetCommit(cacher, dah, int(wrappedTx.ShareIndex), shares.SparseSharesNeeded(uint32(pfb.BlobSize)))
+			commitment, err := inclusion.GetCommit(cacher, dah, int(wrappedTx.ShareIndex), shares.SparseSharesNeeded(pfb.BlobSize))
 			if err != nil {
 				logInvalidPropBlockError(app.Logger(), req.Header, "commitment not found", err)
 				return abci.ResponseProcessProposal{
